Encapsulate logging and printing into a separate class
